- replace calloc by malloc where allocated memory is directly initialized - fix types - optimize round-up division - make local functions "static" - more aggressive memory allocation for realloc()
git-svn-id: https://openmodelica.org/svn/OpenModelica/trunk@15013 f25d12d1-65f4-0310-ae8a-bbce733d8d8e

+++ SimulationRuntime/c/util/integer_array.c
@@ -813,14 +813,14 @@ void range_alloc_integer_array(modelica_integer start, modelica_integer stop, mo
 {
     int n;
 
-    n = floor((stop-start)/inc)+1;
+    n = (int)floor((stop-start)/inc)+1;
     simple_alloc_1d_integer_array(dest,n);
     range_integer_array(start,stop,inc,dest);
 }
 
 void range_integer_array(modelica_integer start, modelica_integer stop, modelica_integer inc, integer_array_t* dest)
 {
-    int i;
+    size_t i;
     /* Assert that dest has correct size */
     for(i = 0; i < dest->dim_size[0]; ++i) {
         integer_set(dest, i, start + (i * inc));
